cleanup java classes and break down code
