+++ app/src/main/java/com/payoneer/payoneerchallenge/network/api/PaymentsService.java
@@ -1,6 +1,5 @@
 package com.payoneer.payoneerchallenge.network.api;
 
-import com.payoneer.payoneerchallenge.network.models.Networks;
 import com.payoneer.payoneerchallenge.network.models.PaymentResponse;
 import io.reactivex.Flowable;
 import retrofit2.http.GET;
