+++ app/src/main/java/com/payoneer/payoneerchallenge/ui/payments/PaymentListAdapter.java
@@ -3,33 +3,16 @@ package com.payoneer.payoneerchallenge.ui.payments;
 import android.view.LayoutInflater;
 import android.view.ViewGroup;
 import androidx.annotation.NonNull;
-import androidx.recyclerview.widget.DiffUtil;
 import androidx.recyclerview.widget.ListAdapter;
 import androidx.recyclerview.widget.RecyclerView;
-import com.bumptech.glide.Glide;
-import com.payoneer.payoneerchallenge.R;
 import com.payoneer.payoneerchallenge.databinding.ItemPaymentMethodBinding;
 import com.payoneer.payoneerchallenge.network.models.ApplicableItem;
+import com.payoneer.payoneerchallenge.utils.AdapterUtils;
 
 public class PaymentListAdapter extends ListAdapter<ApplicableItem, PaymentListAdapter.PaymentListViewHolder> {
 
-    public static final DiffUtil.ItemCallback<ApplicableItem> DIFF_CALLBACK =
-            new DiffUtil.ItemCallback<ApplicableItem>() {
-                @Override
-                public boolean areItemsTheSame(
-                        @NonNull ApplicableItem oldUser, @NonNull ApplicableItem newUser) {
-                    return oldUser.getCode().equals(newUser.getCode());
-                }
-
-                @Override
-                public boolean areContentsTheSame(
-                        @NonNull ApplicableItem oldUser, @NonNull ApplicableItem newUser) {
-                    return oldUser.equals(newUser);
-                }
-            };
-
     public PaymentListAdapter() {
-        super(DIFF_CALLBACK);
+        super(AdapterUtils.DIFF_CALLBACK);
     }
 
     @NonNull
@@ -55,15 +38,7 @@ public class PaymentListAdapter extends ListAdapter<ApplicableItem, PaymentListA
 
         public void bind(ApplicableItem applicableItem) {
             binding.tvPaymentName.setText(applicableItem.getLabel());
-            loadLogo(applicableItem.getLinks().getLogo());
-        }
-
-        private void loadLogo(String imageUrl) {
-            Glide.with(binding.imagePaymentLogo)
-                    .load(imageUrl)
-                    .placeholder(R.drawable.loading_placeholder)
-                    .error(R.drawable.ic_error)
-                    .into(binding.imagePaymentLogo);
+            AdapterUtils.loadLogo(applicableItem.getLinks().getLogo(), binding.imagePaymentLogo);
         }
     }
 }
