+++ app/src/main/java/com/payoneer/payoneerchallenge/utils/Resource.java
@@ -2,9 +2,6 @@ package com.payoneer.payoneerchallenge.utils;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import java.io.IOException;
-import retrofit2.HttpException;
-import timber.log.Timber;
 
 public class Resource<T> {
 
@@ -28,38 +25,11 @@ public class Resource<T> {
     }
 
     public static <T> Resource<T> error(@Nullable String msg, @Nullable T data) {
-        Resource<T> error;
-        if (data instanceof HttpException) {
-            Timber.e("Http exception!");
-            HttpException exception = (HttpException) data;
-            switch (exception.code()) {
-                case 404:
-                    error = new Resource<>(Status.NOT_FOUND, data, msg);
-                    break;
-                case 500:
-                    error = new Resource<>(Status.SERVER_ERROR, data, msg);
-                    break;
-                default:
-                    Timber.e(exception);
-                    error = new Resource<>(Status.UNKNOWN_CODE, data, msg);
-                    break;
-            }
-        } else if (data instanceof IOException) {
-            Timber.e("IO exception!");
-            Timber.e(((IOException) data));
-            error = new Resource<>(Status.N0_CONNECTION, data, msg);
-        } else {
-            Timber.e("Generic exception!");
-            Timber.e("Generic error received %s", data.toString());
-            error = new Resource<>(Status.ERROR, data, msg);
-        }
-        return error;
+        return ResourceUtils.getErrorResource(msg, data);
     }
 
     public static <T> Resource<T> loading() {
         return new Resource<>(Status.LOADING, null, null);
     }
-
-    public enum Status {SUCCESS, ERROR, LOADING, NOT_FOUND, SERVER_ERROR, N0_CONNECTION, UNKNOWN_CODE}
 }
