+++ app/src/main/java/com/payoneer/payoneerchallenge/utils/Status.java
@@ -1,0 +1,4 @@
+package com.payoneer.payoneerchallenge.utils;
+
+public enum Status {SUCCESS, ERROR, LOADING, NOT_FOUND, SERVER_ERROR, N0_CONNECTION, UNKNOWN_CODE}
+
